Add sample python scripts for taking photo
OpenCV is limited to using python2 on the NVIDIA NANO with Raspberry Pi
Camera V2. Python3 can be used with ENV variable
OPENBLAS_CORETYPE=ARMV8 but seems to be unstable.

OpenCV can take a photo or video (at up to 120FPS) and save the photo or
video to a file. The file is in numpy array representation which should
be simple to parse and process into a NN model.