Add basic model and support to GUI
A few things in this commit of note:

1) Fixed the cameraLib to use Python3, turns out this was a dependency
issue on a particular version of OpenCV2 that needed GStreamer Video
support enabled.

2) Updated the GUI to load the model and make predictions. Added some
smarts for generic resizing and model plugging so we can swap models and
resolutions as we need.

3) Added the trained model weights, the model, and a simple test script
to test the model. Most of this is hardcoded to directories on my
machine.

4) Only do one prediction per second, otherwise the camera's FPS drops
way down.